backend/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.2.13
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
